Feature/fix cuda compiler errors and warnings

+++ cudaSift.h
@@ -35,7 +35,7 @@ typedef struct {
 void InitCuda(int devNum = 0);
 float *AllocSiftTempMemory(int width, int height, int numOctaves, bool scaleUp = false);
 void FreeSiftTempMemory(float *memoryTmp);
-void ExtractSift(SiftData &siftData, CudaImage &img, int numOctaves, double initBlur, float thresh, float lowestScale = 0.0f, bool scaleUp = false, float *tempMemory = 0);
+void ExtractSift(SiftData &siftData, CudaImage &img, int numOctaves, float initBlur, float thresh, float lowestScale = 0.0f, bool scaleUp = false, float *tempMemory = 0);
 void InitSiftData(SiftData &data, int num = 1024, bool host = false, bool dev = true);
 void FreeSiftData(SiftData &data);
 void PrintSiftData(SiftData &data);
